[tests] Move tests from chain/client/src/tests to integration-tests directory for consistency
