+++ chain/client/src/lib.rs
@@ -43,7 +43,4 @@ mod stateless_validation;
 pub mod sync;
 pub mod sync_jobs_actor;
 pub mod test_utils;
-
-#[cfg(test)]
-mod tests;
 mod view_client_actor;
