--- chain/client/src/tests/mod.rs
@@ -1,9 +1,0 @@
-mod bug_repro;
-mod catching_up;
-mod chunks_management;
-mod consensus;
-mod cross_shard_tx;
-mod doomslug;
-mod maintenance_windows;
-mod process_blocks;
-mod query_client;
