+++ integration-tests/Cargo.toml
@@ -23,6 +23,7 @@ derive_more.workspace = true
 futures.workspace = true
 hex.workspace = true
 itertools.workspace = true
+num-rational.workspace = true
 parking_lot.workspace = true
 primitive-types.workspace = true
 rand.workspace = true
