+++ integration-tests/src/tests/client/bug_repro.rs
@@ -7,18 +7,15 @@ use std::sync::{Arc, RwLock};
 
 use actix::System;
 use futures::FutureExt;
+use near_actix_test_utils::run_actix;
 use near_async::messaging::CanSend;
 use near_async::time::Clock;
-use near_network::shards_manager::ShardsManagerRequestFromNetwork;
-use near_primitives::types::ShardIndex;
-use rand::{Rng, thread_rng};
-
-use crate::GetBlock;
-use crate::test_utils::{ActorHandlesForTesting, setup_mock_all_validators};
-use near_actix_test_utils::run_actix;
 use near_chain::test_utils::{ValidatorSchedule, account_id_to_shard_id};
+use near_client::GetBlock;
+use near_client::test_utils::{ActorHandlesForTesting, setup_mock_all_validators};
 use near_crypto::InMemorySigner;
 use near_network::client::{BlockApproval, BlockResponse, ProcessTxRequest};
+use near_network::shards_manager::ShardsManagerRequestFromNetwork;
 use near_network::types::NetworkRequests::PartialEncodedChunkMessage;
 use near_network::types::PeerInfo;
 use near_network::types::{
@@ -28,6 +25,8 @@ use near_o11y::WithSpanContextExt;
 use near_o11y::testonly::init_test_logger;
 use near_primitives::block::Block;
 use near_primitives::transaction::SignedTransaction;
+use near_primitives::types::ShardIndex;
+use rand::{Rng, thread_rng};
 
 #[test]
 fn slow_test_repro_1183() {
