+++ integration-tests/src/tests/client/catching_up.rs
@@ -5,13 +5,12 @@ use std::sync::{Arc, RwLock};
 use actix::{Addr, System};
 use borsh::{BorshDeserialize, BorshSerialize};
 use futures::{FutureExt, future};
-use near_async::time::Clock;
-
-use crate::test_utils::{ActorHandlesForTesting, setup_mock_all_validators};
-use crate::{ClientActor, Query};
 use near_actix_test_utils::run_actix;
+use near_async::time::Clock;
 use near_chain::test_utils::{ValidatorSchedule, account_id_to_shard_id};
 use near_chain_configs::TEST_STATE_SYNC_TIMEOUT;
+use near_client::test_utils::{ActorHandlesForTesting, setup_mock_all_validators};
+use near_client::{ClientActor, Query};
 use near_crypto::InMemorySigner;
 use near_network::client::ProcessTxRequest;
 use near_network::types::PeerInfo;
