+++ integration-tests/src/tests/client/process_blocks2.rs
@@ -1,7 +1,7 @@
-use crate::test_utils::TestEnv;
 use assert_matches::assert_matches;
 use near_chain::validate::validate_chunk_with_chunk_extra;
 use near_chain::{Provenance, test_utils};
+use near_client::test_utils::TestEnv;
 use near_crypto::vrf::Value;
 use near_crypto::{KeyType, PublicKey, Signature};
 use near_network::types::{NetworkRequests, PeerManagerMessageRequest};
